refactor(www): own SEO modules directly
