+++ apps/www/app/[locale]/(app)/(shared)/(main)/(learn)/quran/[surah]/page.tsx
@@ -34,12 +34,12 @@ import { recoverStalePublishedQuranSnapshot } from "@/lib/content/quran/recovery
 import { getQuranReferences } from "@/lib/content/quran/references";
 import { VirtualProvider } from "@/lib/context/use-virtual";
 import { getLocaleOrThrow } from "@/lib/i18n/params";
+import { createLocalizedAlternates } from "@/lib/seo/alternates";
+import { createBreadcrumbItems } from "@/lib/seo/breadcrumbs";
+import { generateSEOMetadata } from "@/lib/seo/generator";
+import type { SEOContext } from "@/lib/seo/types";
 import { getSocialMetadata } from "@/lib/utils/metadata";
 import { getQuranPagination, getQuranSurahName } from "@/lib/utils/pages/quran";
-import { createLocalizedAlternates } from "@/lib/utils/seo/alternates";
-import { createBreadcrumbItems } from "@/lib/utils/seo/breadcrumbs";
-import { generateSEOMetadata } from "@/lib/utils/seo/generator";
-import type { SEOContext } from "@/lib/utils/seo/types";
 
 /** Builds localized Quran surah metadata only after the runtime catalog confirms the surah exists. */
 export async function generateMetadata({
