+++ apps/www/lib/seo/types.ts
